examples/fastapi/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.11
sqlalchemy-file==0.6.0
pillow==10.1.0
python-multipart==0.0.6
uvicorn[standard]==0.23.2
